refactor: CardDeck 카드 생성 interface로 분리
