+++ src/main/java/blackjack/domain/card/CardDeck.java
@@ -1,49 +1,15 @@
 package blackjack.domain.card;
 
-import static java.util.stream.Collectors.*;
-
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 import java.util.Stack;
+
+import blackjack.domain.card.deckstrategy.DeckStrategy;
 
 public class CardDeck {
 
     private final Stack<Card> deck;
 
-    private CardDeck(Stack<Card> deck) {
-        this.deck = deck;
-    }
-
-    public static CardDeck create() {
-        List<Card> list = getCards();
-        Collections.shuffle(list);
-
-        Stack<Card> deck = new Stack<>();
-        deck.addAll(list);
-
-        return new CardDeck(deck);
-    }
-
-    public static CardDeck createNoShuffle() {
-        List<Card> list = getCards();
-
-        Stack<Card> deck = new Stack<>();
-        deck.addAll(list);
-
-        return new CardDeck(deck);
-    }
-
-    private static List<Card> getCards() {
-        return Arrays.stream(CardSymbol.values())
-            .flatMap(symbol -> createSymbolCards(symbol).stream())
-            .collect(toList());
-    }
-
-    private static List<Card> createSymbolCards(CardSymbol symbol) {
-        return Arrays.stream(CardNumber.values())
-            .map(cardNumber -> new Card(symbol, cardNumber))
-            .collect(toUnmodifiableList());
+    public CardDeck(DeckStrategy deckStrategy) {
+        this.deck = deckStrategy.create();
     }
 
     public Card drawCard() {
