+++ src/main/java/blackjack/domain/card/deckstrategy/DeckStrategy.java
@@ -1,0 +1,9 @@
+package blackjack.domain.card.deckstrategy;
+
+import java.util.Stack;
+
+import blackjack.domain.card.Card;
+
+public interface DeckStrategy {
+    Stack<Card> create();
+}
