+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import blackjack.domain.NoShuffleDeck;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
 import blackjack.domain.card.CardNumber;
@@ -76,10 +77,11 @@ class PlayerTest {
     @DisplayName("게임준비를 위해 가진 카드를 초기화한다.")
     void init() {
         // give
+        CardDeck cardDeck = new CardDeck(new NoShuffleDeck());
         Player player = new Player("pobi");
 
         // when
-        player.init(CardDeck.createNoShuffle());
+        player.init(cardDeck);
         int actual = player.getScore();
 
         // then
